+++ examples/zod/src/mini/zod/createPetsSchema.ts
@@ -18,6 +18,8 @@ export const createPetsStatusDefaultSchema = petNotFoundSchema
 
 export const createPetsResponseSchema = createPetsStatus201Schema
 
+export const createPetsErrorSchema = createPetsStatusDefaultSchema
+
 export const createPetsDataSchema = z.object({
   name: z.string(),
   tag: z.string(),